pyproject.toml
[tool.poetry]
name = "lightning-hydra-boilerplate"
version = "0.1.0"
description = "A boilerplate for Lightning + Hydra"
authors = ["Willy Fitra Hendria <willyfitrahendria@gmail.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
lightning = "^2.5.0"
hydra-core = "^1.3.2"
torch = "^2.5.1"
omegaconf = "^2.3.0"
optuna = "^4.2.1"
tensorboard = "^2.12.0"
torchvision = "^0.20.0"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
preview = true

# Enable rules
lint.select = [
    "F",    # Pyflakes (`F`)
    "E",    # pycodestyle error (`E`)
    "W",    # pycodestyle warning (`W`)
    "C90",  # mccabe (`C90`)
    "I",    # isort (`I`)
    "N",    # pep8-naming (`N`)
    "D",    # pydocstyle (`D`)
    "UP",   # pyupgrade (`UP`)
    "YTT",  # flake8-2020 (`YTT`)
    "ANN",  # flake8-annotations (`ANN`)
    "S",    # flake8-bandit (`S`)
    "BLE",  # flake8-blind-except (`BLE`)
    "FBT",  # flake8-boolean-trap (`FBT`)
    "B",    # flake8-bugbear (`B`)
    "A",    # flake8-builtins (`A`)
    "COM",  # flake8-commas (`COM`)
    "CPY",  # flake8-copyright (`CPY`)
    "C4",   # flake8-comprehensions (`C4`)
    "DTZ",  # flake8-datatimez (`DTZ`)
    "T10",  # flake8-debugger (`T10`)
    "EM",   # flake8-errmsg (`EM`)
    "FA",   # flake8-future-annotations (`FA`)
    "ISC",  # flake8-implicit-str-concat (`ISC`)
    "ICN",  # flake8-import-conventions (`ICN`)
    "PIE",  # flake8-pie (`PIE`)
    "PT",   # flake8-pytest-style (`PT`)
    "RSE",  # flake8-raise (`RSE`)
    "RET",  # flake8-return (`RET`)
    "SLF",  # flake8-self (`SLF`)
    "SIM",  # flake8-simplify (`SIM`)
    "TID",  # flake8-tidy-imports (`TID`)
    "TCH",  # flake8-type-checking (`TCH`)
    "INT",  # flake8-gettext (`INT`)
    "ARG",  # flake8-unsused-arguments (`ARG`)
    "PTH",  # flake8-use-pathlib (`PTH`)
    "TD",   # flake8-todos (`TD`)
    "FIX",  # flake8-fixme (`FIX`)
    "ERA",  # eradicate (`ERA`)
    "PD",   # pandas-vet (`PD`)
    "PGH",  # pygrep-hooks (`PGH`)
    "PL",   # pylint (`PL`)
    "TRY",  # tryceratos (`TRY`)
    "FLY",  # flynt (`FLY`)
    "NPY",  # NumPy-specific rules (`NPY`)
    "PERF", # Perflint (`PERF`)
    "RUF",  # Ruff-specific rules (`RUF`)
    # "FURB", # refurb (`FURB`) - ERROR: Unknown rule selector: `FURB`
    # "LOG",  # flake8-logging (`LOG`) - ERROR: Unknown rule selector: `LOG`
]

lint.ignore = [
    # pydocstyle
    "D107", # Missing docstring in __init__
    "D415", # First line should end with a period, question mark, or exclamation point

    # pylint
    "PLR0913",  # Too many arguments to function call
    "PLR2004",  # consider replacing with a constant variable
    "PLR0912",  # Too many branches
    "PLR0915",  # Too many statements

    # flake8-annotations
    "ANN101",   # Missing-type-self
    "ANN002",   # Missing type annotation for *args
    "ANN003",   # Missing type annotation for **kwargs

    # flake8-bandit (`S`)
    "S101", # Use of assert detected.

    # flake8-boolean-trap (`FBT`)
    "FBT001",   # Boolean positional arg in function definition
    "FBT002",   # Boolean default value in function definition

    # flake8-datatimez (`DTZ`)
    "DTZ005",    # The use of `datetime.datetime.now()` without `tz` argument is not allowed

    # flake8-fixme (`FIX`)
    "FIX002",    # Line contains TODO, consider resolving the issue
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

# Allow imports relative to the "src" and "tests" directories.
src = ["src", "tests"]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"src/**/*.py" = ["PT"]  # ignore all Pytest rules for source code


[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.mypy]
ignore_missing_imports = true
show_error_codes = true


[[tool.mypy.overrides]]
module = ["torch.*"]
follow_imports = "skip"
follow_imports_for_stubs = true


[tool.pytest.ini_options]
addopts = [
  "--strict-markers",
  "--strict-config",
  "--showlocals",
  "-ra",
  "--cov=src",
  "--cov-report=term-missing",  # Show missing lines in terminal
  "--cov-report=xml",
]
testpaths = "tests"
pythonpath = "src"
markers = [
  "gpu: marks tests that require GPU",
  "cpu: marks tests that can run on CPU only (default)",
]

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']
